Support hotline number encrypt.

+++ aiccs-20191015/src/main/java/com/aliyun/aiccs20191015/models/CreateAgentRequest.java
@@ -4,6 +4,7 @@ package com.aliyun.aiccs20191015.models;
 import com.aliyun.tea.*;
 
 public class CreateAgentRequest extends TeaModel {
+    // js sdk中自动生成的鉴权token
     @NameInMap("ClientToken")
     public String clientToken;
 
@@ -17,10 +18,10 @@ public class CreateAgentRequest extends TeaModel {
     public String displayName;
 
     @NameInMap("SkillGroupId")
-    public java.util.List<Integer> skillGroupId;
+    public java.util.List<Long> skillGroupId;
 
     @NameInMap("SkillGroupIdList")
-    public java.util.List<Integer> skillGroupIdList;
+    public java.util.List<Long> skillGroupIdList;
 
     public static CreateAgentRequest build(java.util.Map<String, ?> map) throws Exception {
         CreateAgentRequest self = new CreateAgentRequest();
@@ -59,19 +60,19 @@ public class CreateAgentRequest extends TeaModel {
         return this.displayName;
     }
 
-    public CreateAgentRequest setSkillGroupId(java.util.List<Integer> skillGroupId) {
+    public CreateAgentRequest setSkillGroupId(java.util.List<Long> skillGroupId) {
         this.skillGroupId = skillGroupId;
         return this;
     }
-    public java.util.List<Integer> getSkillGroupId() {
+    public java.util.List<Long> getSkillGroupId() {
         return this.skillGroupId;
     }
 
-    public CreateAgentRequest setSkillGroupIdList(java.util.List<Integer> skillGroupIdList) {
+    public CreateAgentRequest setSkillGroupIdList(java.util.List<Long> skillGroupIdList) {
         this.skillGroupIdList = skillGroupIdList;
         return this;
     }
-    public java.util.List<Integer> getSkillGroupIdList() {
+    public java.util.List<Long> getSkillGroupIdList() {
         return this.skillGroupIdList;
     }
